requirements.txt
asgiref==3.6.0
Brotli==1.0.9
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
colorama==0.4.6
Django==4.1.5
djangorestframework==3.14.0
filelock==3.9.0
future==0.18.3
gevent==22.10.2
greenlet==2.0.1
gunicorn==20.1.0
idna==3.4
install==1.3.5
joblib==1.2.0
numpy==1.24.1
psycopg2==2.9.5
pycparser==2.21
PySocks==1.7.1
pytz==2022.7.1
requests==2.28.2
requests-file==1.5.1
scikit-learn==1.2.0
scipy==1.10.0
six==1.16.0
sqlparse==0.4.3
threadpoolctl==3.1.0
tldextract==3.4.0
tzdata==2022.7
urllib3==1.26.14
whitenoise==6.3.0
zope.event==4.6
zope.interface==5.5.2
